Factor mocks out
The mocks will also be used for the refresh grant tests.
